+++ frontend/src/views/admin/CmsView.vue
